+++ src/models/operations/index.ts
@@ -11,5 +11,6 @@ export * from "./getorderbyid.js";
 export * from "./getpetbyids.js";
 export * from "./getuserbyname.js";
 export * from "./loginuser.js";
+export * from "./sayhello.js";
 export * from "./updateuser.js";
 export * from "./uploadfile.js";