+++ RELEASES.md
@@ -603,3 +603,23 @@ Based on:
 - [typescript v5.2.0-alpha.23] .
 ### Releases
 - [NPM v5.2.0-alpha.23] https://www.npmjs.com/package/meep-moop/v/5.2.0-alpha.23 - .
+
+## 2025-08-06 11:57:24
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.519.0 (2.552.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v5.2.0-alpha.25] .
+### Releases
+- [NPM v5.2.0-alpha.25] https://www.npmjs.com/package/meep-moop/v/5.2.0-alpha.25 - .
+
+## 2025-08-06 13:26:39
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.519.0 (2.552.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v5.2.0-alpha.27] .
+### Releases
+- [NPM v5.2.0-alpha.27] https://www.npmjs.com/package/meep-moop/v/5.2.0-alpha.27 - .